requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
fonttools==4.44.0
grad-cam==1.5.0
idna==3.4
imageio==2.33.1
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
matplotlib==3.8.1
networkx==3.2.1
numpy==1.26.1
packaging==23.2
pandas==1.3.3
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-metric-learning==2.3.0
pytz==2023.3.post1
requests==2.31.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
threadpoolctl==3.2.0
tifffile==2024.1.30
torch==1.12.1+cu116
torchaudio==0.12.1+cu116
torchvision==0.13.1+cu116
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
